Add function to control the write with threads
For now, the function clk_wiz_init creates a thread. The thread
waits for data from a queue which will contain the offset and the
data for the registets.

A function named queue_send_test has been created to test the
write and read from the queue.

By: Guillermo Gil Rodríguez (guillegilrod@gmail.com)